backend/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==3.1.0
virtualenv
google-auth
